Use src/ to locate the home directory
We had been using the .ark/ directory as a marker for the home
directory. This caused problems with sites stored in git repositories as
git does not sync empty or ignored directories and so attempting to run
`ark build` inside a cloned repository would fail.

We now return to testing for the presence of the src/ directory as a
marker for the home directory.